third_party/WebKit/LayoutTests/fast/dom/shadow/no-style-sharing-with-uncommon-attribute-and-content-combinator-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS window.getComputedStyle(shadow.querySelector("#hostInShadow > .parent > p[isHidden=false]")).display is "block"
PASS window.getComputedStyle(shadow.querySelector("#hostInShadow > .parent > p[isHidden=true]")).display is "none"
